chore(developer): rename to analyzer-messages.ts
Fixes #10254.

A full clean and build resolved this, so seems likely the rename caused
some intermediate file to break?
