+++ developer/src/kmc-analyze/src/analyzer-messages.ts
