+++ developer/src/kmc-analyze/src/index.ts
@@ -5,4 +5,4 @@
 
 export { AnalyzeOskCharacterUse, AnalyzeOskCharacterUseOptions } from './osk-character-use/index.js';
 export { AnalyzeOskRewritePua } from './osk-rewrite-pua/index.js';
-export { AnalyzerMessages } from './messages.js';
+export { AnalyzerMessages } from './analyzer-messages.js';
